Draft implementation of Bolzman mutation
During mutation, check all possible days for a given family and choose
one according to Bolzman distribution with the loss as the energy.

This is terribly slow, though. Check in a draft of a function to compute
losses for all days in a (more) efficient way. That function is not
complete.