Allow poor solutions to survive
The truncation replacement seems decrease the diversity too much. Switch
to selecting survivors for the next generation based on an exponential
distribution instead of a threshold. This way poor solutions can still
survive. Increasing the scale of this distribution is somewhat analogous
to increasing the temperature in simulated annealing.

The old behaviour can be reproduced (approximately) by setting the scale
to a small value, e.g. --survival=0.1.