Implement 2-for-1 swap
Try to move a costly family to a different day and then move two other
families to its former day. There are families with high preference cost
assigned to days with low occupancy. It can happen that swapping such a
family with a different one cannot meet the hard constraints. This
algorithm addresses such a problem.

With this change the rank of the worst choice drops to 4 for the first
time.