More elaborate rule-based improvement
If the best solution hasn't changed, there is no need to try to improve
it again. In that case try to improve some other randomly chosen
solution. In reality, however, the algorithm almost always tries to
improve the best solution because it happens to swap some families
without changing the overall score.

Also alllow running the improvement algorithms for all solutions at the
beginning of the program. This takes a long time doesn't seem to help
really.